+++ app/src/main/java/com/example/physio_plus_app/R3/ParamsR3.java
@@ -1,4 +1,4 @@
-package com.example.physio_plus_app;
+package com.example.physio_plus_app.R3;
 
 public class ParamsR3 {
