+++ app/src/main/java/com/example/physio_plus_app/R3/R3.java
@@ -1,4 +1,4 @@
-package com.example.physio_plus_app;
+package com.example.physio_plus_app.R3;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.Toast;
+
+import com.example.physio_plus_app.R;
 
 public class R3 extends AppCompatActivity {
 
@@ -26,11 +28,11 @@ public class R3 extends AppCompatActivity {
     public void onClick(View v) {
         String namePatient = name_Patient.getText().toString();
         String surnamePatient = surname_Patient.getText().toString();
-        String addressofPatient = addressPatient.getText().toString();
+        String addressOfPatient = addressPatient.getText().toString();
         String amPatient = amkaPatient.getText().toString();
 
         try {
-            String response = OkHttpHandler3.loghistory(new ParamsR3(namePatient,surnamePatient, amPatient, addressofPatient));
+            String response = OkHttpHandler3.loghistory(new ParamsR3(namePatient,surnamePatient, amPatient, addressOfPatient));
             Toast.makeText(R3.this, response, Toast.LENGTH_LONG).show();
         } catch (Exception e) {
             throw new RuntimeException(e);
