+++ app/src/main/res/drawable/amka_r3.xml
