+++ app/src/main/res/drawable/location_r3.xml
