+++ app/src/main/res/drawable/pedio_r3.xml
