+++ app/src/main/res/mipmap-anydpi-v33/ic_launcher.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-    <background android:drawable="@drawable/ic_launcher_background" />
-    <foreground android:drawable="@drawable/ic_launcher_foreground" />
-    <monochrome android:drawable="@drawable/ic_launcher_foreground" />
+    <background android:drawable="@drawable/ic_launcher_background_r3" />
+    <foreground android:drawable="@drawable/ic_launcher_foreground_r3" />
+    <monochrome android:drawable="@drawable/ic_launcher_foreground_r3" />
 </adaptive-icon>
